feat(stdlib): add possibility to fill LEDs with repeatable patterns to the `xod-dev/ws2812` library
