+++ workspace/__lib__/xod-dev/ws2812/color-pattern/patch.cpp
@@ -1,0 +1,15 @@
+{{#global}}
+#include <XColorPattern.h>
+{{/global}}
+
+struct State {
+    Pattern pattern;
+};
+using Type = Pattern*;
+
+{{ GENERATED_CODE }}
+
+void evaluate(Context ctx) {
+    auto state = getState(ctx);
+    emitValue<output_PAT>(ctx, &(state->pattern));
+}
