+++ workspace/__lib__/xod-dev/ws2812/color-pattern/patch.xodp
@@ -1,0 +1,33 @@
+{
+  "description": "Pattern type constructor.",
+  "nodes": [
+    {
+      "id": "BJBsHZ5gI",
+      "position": {
+        "units": "slots",
+        "x": 3,
+        "y": 1
+      },
+      "type": "xod/patch-nodes/utility"
+    },
+    {
+      "id": "HJqn9NzJU",
+      "label": "PAT",
+      "position": {
+        "units": "slots",
+        "x": 1,
+        "y": 2
+      },
+      "type": "xod/patch-nodes/output-self"
+    },
+    {
+      "id": "rkbhqEMy8",
+      "position": {
+        "units": "slots",
+        "x": 1,
+        "y": 1
+      },
+      "type": "xod/patch-nodes/not-implemented-in-xod"
+    }
+  ]
+}
