--- workspace/__lib__/xod-dev/ws2812/example-quickstart/patch.xodp
@@ -1,19 +1,0 @@
-{
-  "nodes": [
-    {
-      "boundLiterals": {
-        "BJZY_6J1L": "#FF0000",
-        "H1wKu611L": "True",
-        "rJ2uOpk1L": "16",
-        "ryD_dp118": "A4"
-      },
-      "id": "BJz7Y6ky8",
-      "position": {
-        "units": "slots",
-        "x": 1,
-        "y": 3
-      },
-      "type": "@/ws2812-mono"
-    }
-  ]
-}
