+++ workspace/__lib__/xod-dev/ws2812/pulse-on-change(color-pattern)/patch.xodp
@@ -1,0 +1,40 @@
+{
+  "nodes": [
+    {
+      "id": "BJoGWJAxU",
+      "position": {
+        "units": "slots",
+        "x": 1,
+        "y": 1
+      },
+      "type": "xod/patch-nodes/utility"
+    },
+    {
+      "id": "HyrhpC6lU",
+      "position": {
+        "units": "slots",
+        "x": 0,
+        "y": 0
+      },
+      "type": "@/input-color-pattern"
+    },
+    {
+      "id": "SJcnT06l8",
+      "position": {
+        "units": "slots",
+        "x": 0,
+        "y": 2
+      },
+      "type": "xod/patch-nodes/output-pulse"
+    },
+    {
+      "id": "SyAnaAagI",
+      "position": {
+        "units": "slots",
+        "x": 0,
+        "y": 1
+      },
+      "type": "xod/patch-nodes/not-implemented-in-xod"
+    }
+  ]
+}
