+++ workspace/__lib__/xod-dev/ws2812/ws2812-device/patch.cpp
@@ -13,7 +13,7 @@ using Type = WS2812*;
 void evaluate(Context ctx) {
     State* state = getState(ctx);
     auto port = getValue<input_PORT>(ctx);
-    uint32_t w = (uint32_t) getValue<input_W>(ctx);
+    uint32_t w = (uint32_t)getValue<input_W>(ctx);
     Type t = new (state->mem) WS2812(port, w);
     emitValue<output_DEV>(ctx, t);
 }
